extract method to determine if character was guessed correctly
